Move executable modules into hades.bin package
