+++ src/hades/config/export.py
@@ -1,16 +1,10 @@
-import argparse
 import collections
 import logging
-import os
 import re
-import sys
 
 import netaddr
 
-from hades.common.cli import ArgumentParser, parser as parent_parser
-from hades.config.loader import load_config
-
-logger = logging.getLogger('hades.config.export')
+logger = logging.getLogger(__name__)
 shell_types = (int, str, bool, netaddr.IPAddress, netaddr.IPNetwork)
 pattern = re.compile(r'([^a-zA-Z0-9_])')
 replacement = r'\\\1'
@@ -51,26 +45,3 @@ def export(config, output_format, file):
             value = ' '.join(escape(v) for v in value
                              if isinstance(v, shell_types))
             print("{}=({})".format(name, value), file=file)
-
-
-def main():
-    parser = ArgumentParser(description='Export options as shell '
-                                                 'variables',
-                            epilog='Python sequence and mapping types will '
-                                   'only be exported, if the destination '
-                                   'format support it',
-                            parents=[parent_parser])
-    parser.add_argument('--format', choices=('systemd', 'posix', 'bash', 'ksh',
-                                             'zsh'),
-                        default='systemd', help='Export format.')
-    parser.add_argument('file', type=argparse.FileType('wb'), metavar='FILE',
-                        default='-', nargs='?',
-                        help='Output destination (default: stdout)')
-    args = parser.parse_args()
-    config = load_config(args.config)
-    export(config, args.format, args.file)
-    return os.EX_OK
-
-
-if __name__ == '__main__':
-    sys.exit(main())
