+++ src/hades/config/generate.py
@@ -3,17 +3,12 @@ import itertools
 import os
 import os.path
 import shutil
-import sys
 from functools import partial
 
 import jinja2
 import netaddr
-import pkg_resources
 from jinja2.exceptions import FilterArgumentError
 from jinja2.filters import environmentfilter
-
-from hades.common.cli import ArgumentParser, parser as common_parser
-from hades.config.loader import load_config
 
 
 def template_filter(name):
@@ -141,34 +136,3 @@ class ConfigGenerator(object):
         stream = self.env.get_template(name).stream(
             BASE_DIRECTORY=base_directory, TARGET=target, **self.config)
         output.writelines(stream)
-
-
-def main():
-    parser = ArgumentParser(parents=[common_parser])
-    parser.add_argument(dest='source', metavar='SOURCE',
-                        help="Template file name or template directory name")
-    parser.add_argument(dest='destination', metavar='DESTINATION', nargs='?',
-                        help="Destination file or directory (default is stdout"
-                             "for files; required for directories)")
-    args = parser.parse_args()
-    config = load_config(args.config)
-    template_dir = pkg_resources.resource_filename('hades.config', 'templates')
-    generator = ConfigGenerator(template_dir, config)
-    source_path = os.path.join(template_dir, args.source)
-    if os.path.isdir(source_path):
-        generator.from_directory(args.source, args.destination)
-    elif os.path.isfile(source_path):
-        if args.destination is None:
-            generator.from_file(args.source, sys.stdout)
-        else:
-            with open(args.destination, 'w', encoding='utf-8') as f:
-                generator.from_file(args.source, f)
-    else:
-        print("No such file or directory {} in {}".format(args.source,
-                                                          template_dir),
-              file=sys.stderr)
-        return os.EX_NOINPUT
-
-
-if __name__ == '__main__':
-    sys.exit(main())
