+++ src/hades/config/templates/uwsgi.ini.j2
@@ -8,7 +8,7 @@ need-app = true
 reaper = true
 chdir = {{ HADES_PORTAL_HOME }}
 die-on-term = false
-module = hades.portal.app
+module = hades.bin.portal
 master = true
 workers = {{ HADES_PORTAL_UWSGI_WORKERS }}
 thunder-lock = true
