+++ systemd/hades-agent.service
@@ -11,6 +11,6 @@ User=hades-agent
 Group=hades-agent
 SyslogIdentifier=hades-agent
 EnvironmentFile=/etc/default/hades
-ExecStart=/usr/local/bin/hades agent
+ExecStart=/usr/local/bin/hades-agent
 Restart=always
 RestartSec=5
