+++ systemd/hades-auth-dns.service
@@ -11,11 +11,11 @@ Type=simple
 PIDFile=/run/hades/auth-dns/unbound.pid
 SyslogIdentifier=hades-auth-dns
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate unbound.conf.j2 /run/hades/auth-dns/unbound.conf
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config unbound.conf.j2 /run/hades/auth-dns/unbound.conf
 ExecStartPre=-/usr/sbin/unbound-anchor -a /var/lib/hades/auth-dns/root.key
 ExecStartPre=/usr/sbin/unbound-checkconf /run/hades/auth-dns/unbound.conf
 ExecStart=/sbin/ip netns exec auth /usr/sbin/unbound -c /run/hades/auth-dns/unbound.conf
-ExecReload=/usr/bin/python3 -m hades.config.generate unbound.conf.j2 /run/hades/auth-dns/unbound.conf
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config unbound.conf.j2 /run/hades/auth-dns/unbound.conf
 ExecReload=/usr/sbin/unbound-control -c /run/hades/auth-dns/unbound.conf reload
 Restart=always
 RestartSec=5
