+++ systemd/hades-auth-vrrp.service
@@ -11,9 +11,9 @@ PIDFile=/run/hades/auth-vrrp/keepalived.pid
 KillMode=process
 SyslogIdentifier=hades-auth-vrrp
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate auth-keepalived.conf.j2 /run/hades/auth-vrrp/keepalived.conf
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config auth-keepalived.conf.j2 /run/hades/auth-vrrp/keepalived.conf
 ExecStart=/sbin/ip netns exec auth /usr/sbin/keepalived --log-console --vrrp --use-file=/run/hades/auth-vrrp/keepalived.conf --pid=/run/hades/auth-vrrp/keepalived.pid --vrrp_pid=/run/hades/auth-vrrp/keepalived-vrrp.pid
-ExecReload=/usr/bin/python3 -m hades.config.generate auth-keepalived.conf.j2 /run/hades/auth-vrrp/keepalived.conf
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config auth-keepalived.conf.j2 /run/hades/auth-vrrp/keepalived.conf
 ExecReload=/bin/kill -HUP $MAINPID
 Restart=always
 RestartSec=5
