+++ systemd/hades-radius-vrrp.service
@@ -11,9 +11,9 @@ PIDFile=/run/hades/radius-vrrp/keepalived.pid
 KillMode=process
 SyslogIdentifier=hades-radius-vrrp
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate radius-keepalived.conf.j2 /run/hades/radius-vrrp/keepalived.conf
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config radius-keepalived.conf.j2 /run/hades/radius-vrrp/keepalived.conf
 ExecStart=/usr/sbin/keepalived --log-console --vrrp --use-file=/run/hades/radius-vrrp/keepalived.conf --pid=/run/hades/radius-vrrp/keepalived.pid --vrrp_pid=/run/hades/radius-vrrp/keepalived-vrrp.pid
-ExecReload=/usr/bin/python3 -m hades.config.generate radius-keepalived.conf.j2 /run/hades/radius-vrpp/keepalived.conf
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config radius-keepalived.conf.j2 /run/hades/radius-vrpp/keepalived.conf
 ExecReload=/bin/kill -HUP $MAINPID
 Restart=always
 RestartSec=5
