+++ systemd/hades-radius.service
@@ -10,10 +10,10 @@ Type=forking
 PIDFile=/run/hades/radius/radiusd.pid
 SyslogIdentifier=hades-radius
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate freeradius /run/hades/radius
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config freeradius /run/hades/radius
 ExecStartPre=/usr/sbin/freeradius -d /run/hades/radius -Cx -lstdout
 ExecStart=/usr/sbin/freeradius -d /run/hades/radius
-ExecReload=/usr/bin/python3 -m hades.config.generate freeradius /run/hades/radius
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config freeradius /run/hades/radius
 ExecReload=/bin/kill -HUP $MAINPID
 Restart=always
 RestartSec=5
