+++ systemd/hades-unauth-http.service
@@ -10,9 +10,9 @@ Type=forking
 PIDFile=/run/hades/unauth-http/nginx.pid
 SyslogIdentifier=hades-unauth-http
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate nginx /run/hades/unauth-http
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config nginx /run/hades/unauth-http
 ExecStart=/sbin/ip netns exec unauth nginx -c /run/hades/unauth-http/nginx.conf
-ExecReload=/usr/bin/python3 -m hades.config.generate nginx /run/hades/unauth-http
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config nginx /run/hades/unauth-http
 ExecReload=/bin/kill -HUP $MAINPID
 Restart=always
 RestartSec=5
