+++ systemd/hades-unauth-portal.service
@@ -9,9 +9,9 @@ Wants=hades-networking.service hades-database.service
 Type=notify
 SyslogIdentifier=hades-unauth-portal
 EnvironmentFile=/etc/default/hades
-ExecStartPre=/usr/bin/python3 -m hades.config.generate uwsgi.ini.j2 /run/hades/unauth-portal/uwsgi.ini
+ExecStartPre=/usr/bin/python3 -m hades.bin.generate_config uwsgi.ini.j2 /run/hades/unauth-portal/uwsgi.ini
 ExecStart=/sbin/ip netns exec unauth /usr/bin/uwsgi --ini=/run/hades/unauth-portal/uwsgi.ini
-ExecReload=/usr/bin/python3 -m hades.config.generate uwsgi.ini.j2 /run/hades/unauth-portal/uwsgi.ini
+ExecReload=/usr/bin/python3 -m hades.bin.generate_config uwsgi.ini.j2 /run/hades/unauth-portal/uwsgi.ini
 ExecReload=/bin/kill -HUP $MAINPID
 Restart=always
 RestartSec=5
